pyproject.toml
[project]
name = "axon"
version = "0.1.0"
description = ""
authors = [
    {name = "Matt Moore",email = "matt@mattmoore.io"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "opentelemetry-api (>=1.30.0,<2.0.0)",
    "opentelemetry-sdk (>=1.30.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.51b0,<0.52)",
    "opentelemetry-instrumentation-requests (>=0.51b0,<0.52)",
    "opentelemetry-opentracing-shim (>=0.51b0,<0.52)",
    "opentelemetry-exporter-otlp (>=1.30.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "axon", from = "src"}]

[tool.poetry.scripts]
start = "axon.routes:start"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
